+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 -----
 
 * Add special case for zero coercion to allow direct use of `sum` method
+* Add useful numeric methods `finite?`, `infinite?`, `zero?`, `nonzero?`, `positive?` and `negative?`
 
 2.8.0
 -----